pyproject.toml
[project]
name = "gsc-api-data"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "google-api-python-client==2.110.0",
    "google-auth==2.25.2",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib==1.2.0",
    "pandas>=2.3.1",
    "psycopg2-binary==2.9.10",
    "python-dotenv==1.0.0",
    "requests==2.31.0",
    "sqlalchemy>=2",
    "uvicorn>=0.35.0",
]
